+++ src/main/java/com/qcloud/Module/Tbaas.java
@@ -1,0 +1,15 @@
+package com.qcloud.Module;
+
+/*
+ * Tencent Blockchain as a Service.
+ * <p>
+ * see API document: <a
+ * href="https://cloud.tencent.com/document/product/663/14024">
+ * https://cloud.tencent.com/document/product/663/14024</a>
+ * </p>
+ */
+public class Tbaas extends Base {
+    public Tbaas(){
+        serverHost = "tbaas.api.qcloud.com";
+    }
+}